app/requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.23.0
psycopg2-binary==2.9.7
redis==4.5.1
elasticsearch==7.10.1
mailgun.py==1.1.3
stripe==5.7.0
pydantic==1.10.8
python-jose==3.3.0
pyjwt==2.6.0
cryptography==41.0.1
fastapi-utils==0.2.1
pandas==2.0.1
xlwt==1.3.0
openpyxl==3.1.2
pdfkit==1.0.0
reportlab==3.6.12
docx==0.2.4
pydantic[email]==1.10.8
aioredis==2.0.1
celery==5.2.7
httpx==0.24.0